Enable client to be run on windows OS.
The motivation behind this change is to be able to use the spotify docker maven plugin on my local dev environment (a windows machine).

CompressedDirectory currently obtains PosixFileAttributes to preserve file permissions when building the tar. When executed on Windows, this throws an UnsupportedOperationException. Have added explicit check to determine if the file system provides a posix compliant view. If so, the existing logic is used to determed the tar file entry mode. Otherwise, a default mode is used.

Initial unit test on the class was failing when run on windows. My change enabled it to pass. I also had to modify the way in which the test instance was constructed because URL.getPath() did not generate a valid windows file path.