Shade/relocate Jersey & its deps to fix conflicts
Jersey and its dependencies (Jackson, JAX-RS, etc.) are now included into
the docker-client JAR and relocated to the com.spotify.docker.client.shaded
package.

This prevents conflicts in projects that use Jersey 1.x. Any project should
be able to use docker-client as a dependency, regardless of which version
of Jersey it uses.

Tested with Helios (Dropwizard 0.7 + Jersey 1.18) and there were no issues
with conflicting Jersey versions after this change.