Add SSL/TLS support
The DockerCertificates class can now be used to build an SSL context for
communicating with Docker over HTTPS with client/server authentication.

Note that we ignore hostnames when verifying SSL certificates. Instead,
we just make sure that the certificate is from the correct CA. Hostname
verification is pointless here since everyone is just using self-signed
certs anyway.

The way we test this stuff is a bit interesting too: we start a container
and run an HTTPS Docker instance inside the container, and see if we can
connect to that with the appropriate certificates.