Disable chunked encoding for container start
Don't send requests to start a container to Docker using chunked encoding.

This is to workaround a bug present in Docker 1.2.0, wherein the HostConfig
is completely ignored if sent with chunked encoding.

To enable testing this, we add ContainerInfo.hostConfig(), which returns
the HostConfig we get from inspecting the container.