Change unrouteable address for testConnectTimeout
Some networks actually try to route the unroutable private IP address that
was being used by testConnectTimeout (172.31.255.1), and then report a
connection refused instead of a timeout, causing the test to fail.

Instead switch to the address 240.0.0.1, which is reserved for future use
and is routable, but is effectively a black hole.