Don't timeout when waiting on a container
The ApacheConnector doesn't respect per-request timeout settings, so the
default read timeout of 30 seconds was being applied when waiting on a
container. This causes an exception on any container that takes longer than
30 seconds to exit (i.e., most containers).

As a workaround, create a second client with infinite read timeout, and use
that for waitContainer and stopContainer. Note that we should actually fix
this bug and send a patch to Jersey for ApacheConnector.

Also add a test, testWaitContainer, to ensure that the correct timeout is
applied.