Add DefaultDockerClient.fromEnv
DefaultDockerClient.fromEnv creates a new DefaultDockerClient.Builder that
is prepopulated with values loaded from DOCKER_HOST and DOCKER_CERT_PATH
environment variables.

This is functionality that was duplicated and spread out among the projects
that use docker-client, so it made sense to just incorporate it into the
client itself, in one place.