CircleCI: Cache public Docker images
Instead of always hitting the public Docker registry, run a mirror locally
that uses S3 in us-east-1 as its backing store.

This should hopefully reduce the number of build failures we get due to
image pulls failing against the public Docker registry. us-east-1 is the
same region that CircleCI workers are in.

Even if there are failures, we now have a log of all communication with the
public Docker registry via the registry mirror log. This should help with
troubleshooting.