Fix broken tests: testDockerDateFormat, testInspectImage
Previously these tests were prone to start failing when the upstream
busybox:latest was changed. The testDockerDateFormat fails as soon as the
busyboxy image is updated, and the testInspectImage fails if the latest version
has no author (which is the case currently). Crappy. Change the test around so
that we pull a busybox image by digest instead -- which means that the "Created"
date on the image will never change, nor will the author or other metadata that
we rely on in these tests.