The JSON created by AuthRegistryConfig is not compatible with Docker api version 1.19+ For version 1.18- it should be {"configs":{"myrepository":{      "username":"myusername",      "password":"mypassword",      "auth":"",      "email":"myemail",      "serveraddress":"myserverAddress"}}}
 For 1.19 it is
 {"auths":{"myrepository":{
      "username":"myusername",
      "password":"mypassword",
      "auth":"",
      "email":"myemail",
      "serveraddress":"myserverAddress"}}}

 For 1.20+ it is
 {"myrepository":{
      "username":"myusername",
      "password":"mypassword",
      "auth":"",
      "email":"myemail",
      "serveraddress":"myserverAddress"}}

Removed the Jackson JSON annotations from AuthRegistryConfig.java and added a @JsonAnyGetter method.
This was done to create the 1.20+ JSON.

Moved versionCompare into its own final class. Made the new compareVersion method static.

In DefaultDockerClient.java's authRegistryHeader method a check was added to figure out what version of api is used, and correct the JSON to match the specific version.
This uses the new compareVersion method.

A test was added to DefaultDockerClientTest.java that checks that the fix is working.
This test uses a newly added Dockerfile.
All tests that used the old versionCompare method now uses the newly created one.