don't use trim on progress streams in AnsiProgressHandler
Trim deletes ANSI escape codes in the stream messages, so you end up
with everything after the escape code and then the message (sans color
or whatever else the escape code was trying to do).

Now the code explicitly just deletes any '\n' chars at the end of the
string as the comment claims it should be doing.