Add integration tests for pushing images
We start a docker registry as a container and push to that.

We don't use @Rule because each method needs a slightly different
startup and teardown.

We don't use @CreateContainer because annotation parameters are decided
at compile-time, but it's hard to know what these parameters should be
at compile-time.

Just do it the old-fashioned way with some code at the beginning and end
of each PushIT method.