Add load/save support
* DockerClient.load allows loading the tarball of an image as a stream. An
image name must be passed in, and the load implementation internally
tags the just loaded image with the provided name. The load
implementation is passed on Docker API's pull, which takes a fromSrc
parameter, indicating it to treat the request's body as the .tar payload
of the image.

* DockerClient.save allows pulling the .tar stream of an image (and saving
it locally or whatever).