Add RANDOM_AVAILABLE_PORT for convenience/clarity.
The API to bind a random available port (as opposed to an explict one) is
currently like PortBinding.of("0.0.0.0", ""), where the empty string (or int 0)
will result in Docker picking a port.

This allows API callers to get the same effect in a more readable fashion using
PortBinding.of("0.0.0.0", PortBinding.RANDOM_AVAILABLE_PORT).