correct API version that cpuQuota was introduced in
Comparing the Docker Remote API docs for [1.19][] versus [1.18][], the
`cpuQuota` parameter was first introduced in 1.19, not 1.18. Update the
test that asserts the version that this is available in.

Spotify's internal Jenkins is using Docker API version 1.18 where this
test fails, which lends more evidence to the idea that it was introduced
in 1.19 and not 1.18.

[1.19]: https://docs.docker.com/engine/reference/api/docker_remote_api_v1.19/
[1.18]: https://docs.docker.com/engine/reference/api/docker_remote_api_v1.18/