tests: only set User argument to ExecCreate when supported
The Docker Remote API documentation does not mention the `User` param to
the `ExecCreate` call at all, so it is not obvious what version it was
added in. `git tag --contains 72a500e9e5929b038816d8bd18d462a19e571c99`
in the docker source code points to `v1.7.0` as the earliest tag that
has it, which is API version 1.19.