update condition on leading slash in testRenameContainer
`docker inspect <container name>` seems to always return a name with a
leading slash, at least as far back as docker version 1.6.2.

So instead of testing the API version to determine which value to test
the container name against, just do an assertion where a leading slash
is ignored if present.