testEventStreamWithSinceTime: filter out events with null status
For some unknown reason, some of the events returned by the docker
events API are returned with null status, which causes the test to fail
because the expected event is not in the correct place in the stream.
This may be isolated to earlier versions of Docker; seems to happen
consistently to us at Spotify with v1.6.2.

Make the test resilient to these types of events by filtering them out.