revert HostConfig.Builder.binds() behavior that appended to list
Partially revert behavior introduced in 1f4d0c7c where the
`HostConfig.Builder.binds(..)` method appends to the existing list
instead of replaces it.

We found a bunch of internal Spotify code that assumed that this method
always did a replacement, and we found it easier to revert the change
here in docker-client (while adding new methods for appending to the
list).