attempt to include HTTP error bodies in exception
When a HTTP error (4xx, 5xx, IOException) occurs in communciating with
the Docker Remote API, DefaultDockerClient attempts to read the response
body (if one exists) and passes it to the DockerRequestException which
is thrown.

The message field that is passed to the DockerRequestException is unused
in the "message" passed up to the superclass (which is available via
`getMessage()` and `Exception.toString()`).

This change would include the "message" passed to DockerRequestException
in the superclass's message if possible, to help improve troubleshooting
and debugging.

Also remove the five unused constructors in DockerRequestException.