use google-auth-library-oauth2-http for fetching access tokens
Instead of invoking `gcloud docker` and parsing a docker config file,
this adds a new ContainerRegistryAuthSupplier which programmatically
fetches access tokens for a Google service or user account loaded for a
credentials file.

The benefit of this implementation is that any arbitrary account can be
used without having to be registered/added to the `gcloud` CLI - the
supplier just needs to be pointed at a file containing the credentials
for the account to be used. There is also no need for gcloud to be
available on the helios agent.

This support should probably be pulled into a new library at some point,
something like `docker-client-gcr`, so that non-GCR users of
docker-client don't have to have this additional logic pulled in. For
now, the Google library needed for GCR support is marked as `optional`
in the pom.xml so users have to make sure to add the dependency
themselves.