remove GoogleContainerRegistryAuthSupplier and Refresher
now that ContainerRegistryAuthSupplier is able to fetch access tokens
for arbitrary accounts and the default application credentials, the
original implementation that invokes `gcloud docker -a` and reads the
configuration file seems unnecessary. In every case it would be better
to use the new ContainerRegistryAuthSupplier, rather than requiring
`gcloud` to be installed and having it populate a json file used by the
docker CLI which we then read.