default to pulling auth info from docker config file
This commit changes the behavior of DefaultDockerClient to default to
using authentication info from the docker-cli configuration file. This
will allow users to get the same authentication support they would
normally have from the docker-cli out-of-the-box.

Additionally, MultiRegistryAuthSupplier can be used to combine multiple
RegistryAuthSuppliers - for instance, to be able to combine the dynamic
"fetch AccessTokens" support for Google Container Registry with the
ability to read authentication info from the config file for non-GCR
registries.