Support loading auth config for repos with no scheme in their URL
Support loading auth configuration for repos when their `serverAddress`
is specified with no scheme component but their entry in the configuration
file includes a scheme component.

Older versions of Docker write authentication tokens to the configuration
file including their scheme ("https://") but newer onces do not. Thus
in order to allow people to use the newer format (looking up auth for
a repo that isn't prefixed with scheme) with an older configuration file
we try prefixing the `serverAddress` with "http://" in some cases.

Fixes #804