Fix a NullPointerException with some credHelpers, like GCE
## Symptoms

When docker is configured to authenticate to a private Google Container Registry (GCR for short), a credHelper of type `gcr` is added to `.docker/config.json`.

This renders this library unusable.
Simply running this code snippet:

```java
dockerClient = DefaultDockerClient.fromEnv().build();
dockerClient.build("/tmp", "superimage");
```

Throws the following NPE:

```
java.lang.NullPointerException: null value in entry: appengine.gcr.io=null
        at com.google.common.collect.CollectPreconditions.checkEntryNotNull(CollectPreconditions.java:34) ~[guava-20.0.jar:na]
        at com.google.common.collect.ImmutableMapEntry.<init>(ImmutableMapEntry.java:49) ~[guava-20.0.jar:na]
        at com.google.common.collect.ImmutableMap.entryOf(ImmutableMap.java:122) ~[guava-20.0.jar:na]
        at com.google.common.collect.ImmutableMap$Builder.put(ImmutableMap.java:198) ~[guava-20.0.jar:na]
        at com.spotify.docker.client.messages.RegistryConfigs$Builder.addConfig(RegistryConfigs.java:103) ~[docker-client-8.13.1.jar:8.13.1]
        at com.spotify.docker.client.DockerConfigReader.authForAllRegistries(DockerConfigReader.java:135) ~[docker-client-8.13.1.jar:8.13.1]
        at com.spotify.docker.client.auth.ConfigFileRegistryAuthSupplier.authForBuild(ConfigFileRegistryAuthSupplier.java:106) ~[docker-client-8.13.1.jar:8.13.1]
        at com.spotify.docker.client.DefaultDockerClient.build(DefaultDockerClient.java:1468) ~[docker-client-8.13.1.jar:8.13.1]
        at com.spotify.docker.client.DefaultDockerClient.build(DefaultDockerClient.java:1445) ~[docker-client-8.13.1.jar:8.13.1]
        at com.spotify.docker.client.DefaultDockerClient.build(DefaultDockerClient.java:1431) ~[docker-client-8.13.1.jar:8.13.1]
```

## Cause

`DockerConfigReader` tries to authenticate for all registries, including the GCR ones.

To do this, it repeatedly executes the command:

`docker-credential-${credHelper} get`

feeding it the registry url via stdin.

If the executed command returns `credentials not found in native keychain`, `null` is returned and later gets stored as a value in a Guava `ImmutableMap`, which throws the `NPE`.

The GCR credsHelper works differently, as it requires you to execute:

`docker-credential-gcr gcr-login`

Which opens a browser with the Google auth page.

## Proposed fix

If the authentication to a registry fails, skip it and do not try to store `null` as a value in the authentications map.