Fixed bug in a unit test which caused test to block indefinitely
The test DefaultDockerClientTest.testBuildInterruption() was lacking an
Exception catch-all. Without it, the spawned thread would not complete
the 'started' future and the test would block forever waiting for it to
complete.

This would happen on Windows systems due to a bug also fixed by this
commit. The code was using URL.getPath() to initialize a Path. This
is not safe on Windows systems due to the drive letter (eg 'C:'). The
fix uses standard NIO APIs to produce a Path from the URL. This is
done in a platform agnostic manner.